Move NAV constants to top level
This was only needed locally in Groups but we can to store all  nav constants in the global constants file
