+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/constants.ts
@@ -3,6 +3,23 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
+import { i18n } from '@kbn/i18n';
+
+export const NAV = {
+  GROUPS: i18n.translate('xpack.enterpriseSearch.workplaceSearch.nav.groups', {
+    defaultMessage: 'Groups',
+  }),
+  GROUP_OVERVIEW: i18n.translate(
+    'xpack.enterpriseSearch.workplaceSearch.nav.groups.groupOverview',
+    {
+      defaultMessage: 'Overview',
+    }
+  ),
+  SOURCE_PRIORITIZATION: i18n.translate(
+    'xpack.enterpriseSearch.workplaceSearch.nav.groups.sourcePrioritization',
+    { defaultMessage: 'Source Prioritization' }
+  ),
+};
 
 export const MAX_TABLE_ROW_ICONS = 3;
